# Conflicts:
#	epe-backend/src/main/resources/application.properties

+++ epe-backend/src/main/java/com/isdintership/epe/dto/EvaluationGroupDto.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Data
 public class EvaluationGroupDto {
     private String id;
-    private Integer overallScore;
+    private Float overallScore;
     private String title;
     private String assessmentId;
     @JsonProperty("evaluationFields")
